backend/requirements.txt
uagents>=0.13.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
google-genai>=1.0.0
pymongo[srv]>=4.6.0
elevenlabs>=1.0.0
python-dotenv>=1.0.0
websockets>=12.0
pydantic>=2.0.0
httpx>=0.27.0
ytmusicapi>=1.7.0
